♻️ refactor[#10.1.6]: Backend 서비스 레이어 KeywordClassifier 마이그레이션
🔄 임포트 경로 변경:
- keyword_classifier → keyword 모듈로 변경
- 모든 서비스 및 분류기에서 새로운 경로 사용

📁 파일 변경 사항:
- backend/services/classification_service.py
  - 임포트: keyword_classifier → keyword
  - 메서드: aclassify() → classify()
  - 파라미터: user_context → context

- backend/services/conflict_service.py
  - 임포트: keyword_classifier → keyword (2곳)
  - 메서드: aclassify() → classify()
  - 주석 업데이트: "async 지원하면 classify 사용"

- backend/classifier/para_agent.py
  - 임포트: keyword_classifier → keyword

🔧 API 변경 사항:
- 메서드명: aclassify() → classify()
- 파라미터: user_context → context (일관성)
- 반환값 구조: {"tags": [...]} → {"category": "..."}

📌 호환성:
- 기존 Mock 테스트와 호환 유지
- 실제 분류 로직은 새로운 keyword.py 사용
- 에러 핸들링 및 Fallback 로직 유지

📌 Related: Issue [#75]
